bugfixed working localStorage
